+++ lib/streamlit/runtime/app_session.py
@@ -19,9 +19,6 @@ from asyncio import AbstractEventLoop
 from enum import Enum
 from typing import TYPE_CHECKING, Callable, Dict, Optional, List, Union
 
-from streamlit.proto.BackMsg_pb2 import BackMsg
-from streamlit.uploaded_file_manager import UploadedFileManager
-
 import streamlit.elements.exception as exception_utils
 from streamlit import __version__, caching, config, legacy_caching, secrets, source_util
 from streamlit.case_converters import to_snake_case
@@ -29,6 +26,7 @@ from streamlit.credentials import Credentials
 from streamlit.in_memory_file_manager import in_memory_file_manager
 from streamlit.logger import get_logger
 from streamlit.metrics_util import Installation
+from streamlit.proto.BackMsg_pb2 import BackMsg
 from streamlit.proto.ClientState_pb2 import ClientState
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
 from streamlit.proto.GitInfo_pb2 import GitInfo
@@ -39,13 +37,14 @@ from streamlit.proto.NewSession_pb2 import (
     UserInfo,
 )
 from streamlit.proto.PagesChanged_pb2 import PagesChanged
-from streamlit.session_data import SessionData
-from streamlit.runtime.scriptrunner import (
+from streamlit.uploaded_file_manager import UploadedFileManager
+from streamlit.watcher import LocalSourcesWatcher
+from .scriptrunner import (
     RerunData,
     ScriptRunner,
     ScriptRunnerEvent,
 )
-from streamlit.watcher import LocalSourcesWatcher
+from .session_data import SessionData
 
 LOGGER = get_logger(__name__)
 if TYPE_CHECKING:
